requirements.txt
fastapi
uvicorn
langchain
langchain-openai
pydantic
requests
psutil
python-dotenv
streamlit
